feat: add more graceful error handling if deleting files fails

+++ app/components/manager/FileDisplay.tsx
@@ -166,24 +166,29 @@ const FileDisplay:React.FC<FileDisplayProps> = ({ subCorpusFiles }) => {
     /**
      * Deleting files that are no longer needed
      */
-    const handleDeleteFile = (file_id: number) => {
-        // TODO
+    const handleDeleteFile = async (file_id: number) => {
         alert("This file will be deleted!");
-        const result = deleteFile(file_id);
-        corpusDispatch({ type: 'delete-file', subCorpusId: subCorpusFiles.subCorpus.id, fileId: file_id });
+        const result = await deleteFile(file_id);
+        if (result.status === 'success') {
+            corpusDispatch({ type: 'delete-file', subCorpusId: subCorpusFiles.subCorpus.id, fileId: file_id });
+            toast.success("File successfully deleted.");
+        } else {
+            toast.error("There was an error deleting your file: " + result.cppOutput);
+        }
     }
 
     /**
      * Functionality to delete the entire subcorpus
      */
-    const handleDeleteSubcorpus = () => {
+    const handleDeleteSubcorpus = async () => {
         alert("Are you sure you want to do this? It can't be undone!");
-        const result = deleteSubcorpus(subCorpusFiles.subCorpus.id);
-        // Handle the result here
-        // If the subcorpus is deleted, show a message that it was deleted
-        // May need a prop to message up to parent Manager.tsx to update view
-        // and remove this component from the view.
-        corpusDispatch( { type: 'delete-subcorpus', subCorpusId: subCorpusFiles.subCorpus.id });
+        const result = await deleteSubcorpus(subCorpusFiles.subCorpus.id);
+        if (result.status === 'success') {
+            corpusDispatch( { type: 'delete-subcorpus', subCorpusId: subCorpusFiles.subCorpus.id });
+            toast.success("Subcorpus group successfully deleted.");
+        } else {
+            toast.error("There was an error deleting the subcorpus: " + result.cppOutput);
+        }
     }
 
     return (
